+++ clang-tools-extra/clang-tidy/bugprone/PosixReturnCheck.cpp
@@ -28,26 +28,31 @@ static StringRef getFunctionSpelling(const MatchFinder::MatchResult &Result,
 }
 
 void PosixReturnCheck::registerMatchers(MatchFinder *Finder) {
-  Finder->addMatcher(binaryOperator(hasOperatorName("<"),
-                                    hasLHS(callExpr(callee(functionDecl(
-                                        matchesName("^::posix_"),
-                                        unless(hasName("::posix_openpt")))))),
-                                    hasRHS(integerLiteral(equals(0))))
-                         .bind("ltzop"),
-                     this);
-  Finder->addMatcher(binaryOperator(hasOperatorName(">="),
-                                    hasLHS(callExpr(callee(functionDecl(
-                                        matchesName("^::posix_"),
-                                        unless(hasName("::posix_openpt")))))),
-                                    hasRHS(integerLiteral(equals(0))))
-                         .bind("atop"),
-                     this);
+  Finder->addMatcher(
+      binaryOperator(
+          hasOperatorName("<"),
+          hasLHS(callExpr(callee(functionDecl(
+              anyOf(matchesName("^::posix_"), matchesName("^::pthread_")),
+              unless(hasName("::posix_openpt")))))),
+          hasRHS(integerLiteral(equals(0))))
+          .bind("ltzop"),
+      this);
+  Finder->addMatcher(
+      binaryOperator(
+          hasOperatorName(">="),
+          hasLHS(callExpr(callee(functionDecl(
+              anyOf(matchesName("^::posix_"), matchesName("^::pthread_")),
+              unless(hasName("::posix_openpt")))))),
+          hasRHS(integerLiteral(equals(0))))
+          .bind("atop"),
+      this);
   Finder->addMatcher(
       binaryOperator(
           anyOf(hasOperatorName("=="), hasOperatorName("!="),
                 hasOperatorName("<="), hasOperatorName("<")),
           hasLHS(callExpr(callee(functionDecl(
-              matchesName("^::posix_"), unless(hasName("::posix_openpt")))))),
+              anyOf(matchesName("^::posix_"), matchesName("^::pthread_")),
+              unless(hasName("::posix_openpt")))))),
           hasRHS(unaryOperator(hasOperatorName("-"),
                                hasUnaryOperand(integerLiteral()))))
           .bind("binop"),